Moved stories to component folders
