+++ .storybook/config.js
@@ -43,9 +43,9 @@ addParameters({
 });
 
 // automatically import all files ending in *.stories.js
-const req = require.context('../stories', true, /\.stories\.js$/);
-function loadStories() {
-  req.keys().forEach(filename => req(filename));
-}
+const req = require.context('../src', true, /stories\.js$/);
 
-configure(loadStories, module);
+configure(
+  () => req.keys().forEach(filename => req(filename)),
+  module
+);
