+++ src/Field/stories.js
@@ -1,0 +1,51 @@
+import React from 'react'
+
+import create from '../create'
+import { action } from '@storybook/addon-actions'
+
+import Field from './'
+import FieldDescription from './FieldDescription'
+import Input from '../Input'
+
+create('Field')
+  .add('Default', () => (
+    <Field
+      id="field-test"
+      label="Test"
+      description="Test description"
+      hint="(optional)">
+      <Input placeholder="Text placeholder" />
+    </Field>
+  ))
+  .add('Error', () => (
+    <Field
+      id="field-error"
+      error
+      label="Test"
+      description="Test description"
+      hint="(optional)">
+      <Input placeholder="Text placeholder" />
+    </Field>
+  ))
+  .add('Disabled', () => (
+    <Field
+      id="field-disabled"
+      disabled
+      label="Test"
+      description="Test description"
+      hint="(optional)">
+      <Input placeholder="Text placeholder" />
+    </Field>
+  ))
+  .add('Custom description', () => (
+    <figure>
+      <Field
+        id="field-cutsom-description"
+        label="Test"
+        customDescription={<FieldDescription>Test custom content with <a href="#" onClick={action('onClick event')}>link</a></FieldDescription>}
+        hint="(optional)">
+        <Input placeholder="Text placeholder" />
+      </Field>
+    </figure>
+  ))
+
