+++ src/Field/test.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import Field from './'
+import FieldDescription from './FieldDescription'
 import Input from '../Input'
 import renderer from 'react-test-renderer'
 
@@ -41,6 +42,20 @@ it('Field with error', () => {
 it('Field with custom description with error', () => {
   const tree = renderer
     .create(<Field id="test" label="Test label" error={true} customDescription={<span>Custom description</span>}><Input /></Field>)
+    .toJSON()
+  expect(tree).toMatchSnapshot()
+})
+
+it('Default FieldDescription', () => {
+  const tree = renderer
+    .create(<FieldDescription>Description</FieldDescription>)
+    .toJSON()
+  expect(tree).toMatchSnapshot()
+})
+
+it('FieldDescription with error', () => {
+  const tree = renderer
+    .create(<FieldDescription error={true}>Description</FieldDescription>)
     .toJSON()
   expect(tree).toMatchSnapshot()
 })
