+++ src/ProgressCircle/stories.js
@@ -1,16 +1,17 @@
 import React from 'react'
 
-import create from './create'
+import create from '../create'
 import { number, radios, boolean } from '@storybook/addon-knobs'
 
-import { ProgressCircle } from '../dist'
+import ProgressCircle from './'
 
-create('Progress Circle')
-  .add('Default', () => (
-    <div>
+create('ProgressCircle')
+  .add('Viewer', () => (
+    <figure>
       <ProgressCircle
         variant={radios('Variant', ['null', 'error', 'success', 'warning'])}
         value={number('Value', 20)}
         busy={boolean('Busy', false)} />
-    </div>
+    </figure>
   ))
+
