+++ src/index.css
@@ -1,6 +1,6 @@
 @import './variables.css';
 @import './defaults.css';
-@import './utils.css';
+@import './utils/utils.css';
 
 /* Components */
 @import './Button/style.css';
